requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
jinja2>=3.1.0
pytz>=2024.1
tweepy>=4.14.0
Pillow>=10.0.0
yfinance>=0.2.36
flask>=3.0.0
stripe>=7.0.0
